Debug and test improvements
git-svn-id: file:///home/tiste/MOJOHAUS-TO-GIT/SVN-MOJO-WIP/trunk/mojo/findbugs-maven-plugin@14765 52ab4f32-60fc-0310-b215-8acea882cd1b

+++ src/main/groovy/org/codehaus/mojo/findbugs/FindBugsMojo.groovy
@@ -674,10 +674,14 @@ class FindBugsMojo extends AbstractMavenReport implements FindBugsInfo {
     List reportPlugins = getProject().getReportPlugins()
 
     reportPlugins.each() {reportPlugin ->
+      
+      log.info("report plugin -> ${reportPlugin.getArtifactId()}")
       if ( "maven-jxr-plugin".equals(reportPlugin.getArtifactId()) || "jxr-maven-plugin".equals(reportPlugin.getArtifactId()) ) {
         isEnabled = true
       }
     }
+    
+    log.info("jxr report links are ${isEnabled?"enabled":"disabled"}")
     return isEnabled
   }
 
@@ -950,8 +954,11 @@ class FindBugsMojo extends AbstractMavenReport implements FindBugsInfo {
       } else {
         log.info("No bugs found")
       }
+      
+      if (!log.isDebugEnabled()) {
+        tempFile.delete()
+      }
 
-      tempFile.delete()
     }
 
     if (outputFile.exists()) {
